+++ packages/terra-date-time-picker/src/DateTimePicker.jsx
@@ -69,6 +69,10 @@ const propTypes = {
    */
   onChangeRaw: PropTypes.func,
   /**
+   * A callback function triggered when the date input, hour input, or minute input receives focus. function(event)
+   */
+  onFocus: PropTypes.func,
+  /**
    * A callback function to let the containing component (e.g. modal) to regain focus.
    */
   releaseFocus: PropTypes.func,
@@ -99,6 +103,7 @@ const defaultProps = {
   onBlur: undefined,
   onChange: undefined,
   onChangeRaw: undefined,
+  onFocus: undefined,
   releaseFocus: undefined,
   requestFocus: undefined,
   timeInputAttributes: undefined,
@@ -313,6 +318,10 @@ class DateTimePicker extends React.Component {
   }
 
   handleOnInputFocus(event) {
+    if (this.props.onFocus) {
+      this.props.onFocus(event);
+    }
+
     if (!this.isDefaultDateAcceptable) {
       this.dateValue = '';
       this.timeValue = '';
@@ -427,6 +436,7 @@ class DateTimePicker extends React.Component {
       onBlur,
       onChange,
       onChangeRaw,
+      onFocus,
       maxDateTime,
       minDateTime,
       name,
@@ -462,7 +472,7 @@ class DateTimePicker extends React.Component {
           onSelect={this.handleOnSelect}
           onClickOutside={this.handleOnClickOutside}
           onBlur={this.handleOnDateBlur}
-          onInputFocus={this.handleOnDateInputFocus}
+          onFocus={this.handleOnDateInputFocus}
           excludeDates={excludeDates}
           filterDate={filterDate}
           includeDates={includeDates}
@@ -480,7 +490,7 @@ class DateTimePicker extends React.Component {
           <TimeInput
             onBlur={this.handleOnTimeBlur}
             onChange={this.handleTimeChange}
-            onInputFocus={this.handleOnTimeInputFocus}
+            onFocus={this.handleOnTimeInputFocus}
             inputAttributes={timeInputAttributes}
             name="input"
             value={this.timeValue}