+++ packages/terra-time-input/CHANGELOG.md
@@ -3,8 +3,14 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added 'onFocus' prop.
+
 ### Changed
 * Change the onBlur event to get triggered only when the entire time input component loses focus.
+
+### Removed
+* Removed the 'onInputFocus' custom prop in favor of the first-class 'onFocus' prop.
 
 3.6.0 - (March 26, 2019)
 ------------------